# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ZGalaxy - Planet & Moon Infrastructure Management API
  description: |
    ZGalaxy Planet & Moon Infrastructure Engine API. Manages self-hosted ZeroTier
    Planet/Moon infrastructure: authentication & RBAC, planet/moon building,
    federation mesh, cluster HA, Cloudflare DNS/DDNS, identity, and encrypted backup.
  version: 1.3.0
servers:
  - url: http://localhost:3000
    description: Local ZGalaxy server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
    Session:
      type: object
      properties:
        token:
          type: string
        username:
          type: string
        role:
          type: string
          enum: [ADMIN, OPERATOR, READ_ONLY]
        createdAt:
          type: string
          format: date-time
paths:
...
  /api/v1/auth/login:
    post:
      summary: Authenticate and obtain a bearer session token (public)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        '200':
          description: Session token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '401':
          $ref: '#/components/schemas/Error'
        '429':
          description: Rate limited
  /api/v1/auth/logout:
    post:
      summary: Invalidate the current session token
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logged out
  /api/v1/auth/me:
    get:
      summary: Current authenticated user profile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Profile
  /api/v1/auth/users:
    get:
      summary: List users (ADMIN only)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of users without hashes
        '403':
          $ref: '#/components/schemas/Error'
  /api/v1/auth/users/create:
    post:
      summary: Create a user (ADMIN only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password, role]
              properties:
                username: { type: string }
                password:
                  type: string
                  minLength: 8
                role:
                  type: string
                  enum: [ADMIN, OPERATOR, READ_ONLY]
      responses:
        '201':
          description: User created
        '400':
          $ref: '#/components/schemas/Error'
        '403':
          $ref: '#/components/schemas/Error'
  /api/v1/auth/users/{username}:
    delete:
      summary: Delete a user (ADMIN only; `admin` is protected)
      security:
        - BearerAuth: []
      parameters:
        - name: username
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: User deleted
        '400':
          $ref: '#/components/schemas/Error'
        '403':
          $ref: '#/components/schemas/Error'

  # ================= PLANET =================
...
  /api/v1/planet:
    delete:
      summary: Delete the planet file (unlinks; ADMIN only)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Planet deleted
        '403':
          $ref: '#/components/schemas/Error'
...
  /api/v1/federation/handshake:
    post:
      summary: Inter-node handshake (public; token-gated, rate limited, anti-poisoning)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sourceNodeId, sourceEndpoint, tokenSecret]
              properties:
                sourceNodeId: { type: string }
                sourceNodeName: { type: string }
                sourceEndpoint: { type: string }
                tokenSecret: { type: string }
                requestedSyncMode:
                  type: string
                  enum: [FEDERATION_INHERITED, DIRECT_ISOLATED]
      responses:
        '200':
          description: Handshake accepted
        '400':
          $ref: '#/components/schemas/Error'
        '429':
          description: Rate limited
...
  /api/v1/federation/join:
    post:
      summary: Join a remote federation node via token (SSRF-guarded)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [targetEndpoint, tokenSecret]
              properties:
                targetEndpoint: { type: string }
                tokenSecret: { type: string }
                syncMode:
                  type: string
                  enum: [FEDERATION_INHERITED, DIRECT_ISOLATED]
      responses:
        '200':
          description: Joined
        '400':
          description: Invalid/blocked target (SSRF protection)
        '403':
          $ref: '#/components/schemas/Error'
...
  /api/v1/cloudflare/config:
    get:
      summary: Get Cloudflare configuration (apiToken masked)
      responses:
        '200':
          description: Config with apiTokenMasked / hasApiToken
    post:
      summary: Save Cloudflare configuration (ADMIN/OPERATOR)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                enabled: { type: boolean }
                mode: { type: string }
                apiToken: { type: string }
                zoneId: { type: string }
                zoneName: { type: string }
                recordId: { type: string }
                recordName: { type: string }
                recordType: { type: string }
                proxied: { type: boolean }
                autoRebuildPlanet: { type: boolean }
      responses:
        '200':
          description: Config saved (masked)
        '403':
          $ref: '#/components/schemas/Error'
  /api/v1/cloudflare/verify-token:
    post:
      summary: Verify a Cloudflare API token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                apiToken: { type: string }
      responses:
        '200':
          description: Validation result
...
  /api/v1/cloudflare/zones/{zoneId}/records:
    get:
      summary: List DNS records for a zone
      parameters:
        - name: zoneId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Records
  /api/v1/cloudflare/sync:
    post:
      summary: Trigger manual Cloudflare DNS synchronization (ADMIN/OPERATOR)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Sync result
        '403':
          $ref: '#/components/schemas/Error'
  /api/v1/cloudflare/logs:
    get:
      summary: Cloudflare sync history
      responses:
        '200':
          description: Logs
    delete:
      summary: Clear Cloudflare sync logs (ADMIN/OPERATOR)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logs cleared
        '403':
          $ref: '#/components/schemas/Error'

  # ================= NETWORK =================
...